Version 2021.1.1 (06.05) (#15)
* Add go.mod and database module

* Add user model and register request.

* Add hash and password comparing functions.

* Update register user SQL query.

* Database query executing error handling.

* Some fixes to error handling, when user registered.

* Add firstName and lastName fields to User model.

* Add handling username and lastname on registering to SQL query.

* Add sign up and access token generation.

* Update Dockerfile

* Add validation on-register.

* Update gomod.

* Add 'this username already taken' error.

* Add CORS policy.

* Fix typo in CORS policy at main.go

* Update go.yml

* Add README.md.

* Features/100 database api (#4)

* #98 Add login

* #90 Basic inits

* #105 Add add endpoint

* #107 Add delete endpoint

* #98 edit VerifyUser, #109 add hide endpoint, #106 add edit endpoint

* #109 Add hide endpoint

* #123 Add suggestion endpoint

* #110 Add recieve endpoint

* #105 Add endpoint edit

* #108 1136 error fix

* #108 1292 error fix

* #110 endpoint fix

* #98 fix comparison between pass and hash

* #120 Add authorization control

* Edit error catch at row 27 in main.go

* Merge conflict complete

* Add valid access token return on succesful login.

* Change server port for production build.

* Cleanup code, check token for nil.

* Reformat edit to update

* Add error displaying on API request.

* Add error handlings and GetWishes id argument.

* changes for readable code

* Error handling.

* #119 Add follow endpoints

* Invalid credentials error handling.

* follow endpoints implementation fix

* #126 refactor username to email

* Server error change to UnAuthorization error

* Fix some bugs.


* Fix recieve GET query.

* Move receive endpoint to user related data.

* Fix typo in get user ID query.

* Update wishes structure to display users' information in wishes.

* Make /user/ endpoint POST -> GET

* Update /wish/add structure

* Update wish model

* Add creationTime on rows.Scan() for GetWishes

* Update CreationTime type for SQL compatibility.

* Concat creationTime to string.

* Add profile getting on /user/:id

* Hide MySQL connection string and update README.md

* Cleanup .yml and README

* Delete log.fatal and catch no rows.

* Fix error message for user not found in DB.

* Delete log.fatal() and take DB fixes.

* Fix go-version in matrix for GitHub Actions

* Fix go.yml.

* Add /POST/ user endpoint for editing current user's profile (#6)

* Add POST /user/ endpoint to update current user's profile.

* Fix data assignment on editing user profile.

* Add checking, that new user first name is not null.

* Fix go.yml.

* Add deleting wish feature with permissions check and migrate to new naming convention. (#7)

* Fix go.yml.

* Fix wishes deleting endpoint and rename to new naming convention.

* Profile edit validation requires JSON validation & error catching. (#8)

* Fix go.yml.

* Hotfix invalid request body on invalid JSON for editing profile.

* Follow trackings, when requesting user's page. (#9)

* Fix go.yml.

* Add tracking following on GET /user/:id endpoint.

* Add check is followed module.

* #119 - Add toggled followed and unfollowed. (#10)

* #144 - A lot of places in code uses different types of int. (#11)

* Add return for handled error, when adding wish.

* Add return in add function, when deleting wish.

* Add return to login function.

* Fix error rewriting in profile loading

* Add return statement on wish hiding error

* Add return statement on wishes suggestion receiving error

* Add return statement on wishes hiding error

* #147, #139 -  Migrate adding wishes to PUT method and add permissions validation for actions, releated with wishes (#12)

* #147 - Auto-assign wish to current user and migrate to PUT method

* #147 - Fix permissions in actions, releated with wishes

* #131 - Refactor endpoints (#14)

+++ models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID               uint64 `json:"id,omitempty"`
+	ID               int    `json:"id,omitempty"`
 	FirstName        string `json:"firstName,omitempty"`
 	LastName         string `json:"lastName,omitempty"`
 	Email            string `json:"email,omitempty"`
